finishing build

+++ app/Http/Controllers/ArticlesController.php
@@ -6,15 +6,29 @@ use Illuminate\Http\Request;
 use App\Article;
 use App\Category;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Storage;
 
 class ArticlesController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware('auth',['except' => ['index','show']]);
+    }
+
     public function index()
     {
         $articles = Article::orderBy('created_at','desc')->paginate(5);
         return view('articles.index')->with('articles',$articles);
     }
 
+    public function categoriesArticles($name){
+        $category = Category::where('name',$name)->first();
+        // dd($category->id);
+        $articles = Article::where('category_id', $category->id)->paginate(5);;
+        // dd($articles);
+        return view('articles.index')->with('articles',$articles);
+    }
 
     public function create()
     {
@@ -29,9 +43,25 @@ class ArticlesController extends Controller
             'title' => 'required',
             'desc' => 'required',
             'cat' => 'required',
+            'image' => 'image|nullable|max:1999'
         ]);
 
-        // dd( request()->all() );
+        // handle file
+        if($request->hasFile('image')){
+            // Get filename with extension
+            $fileNameWithExt = $request->file('image')->getClientOriginalName();
+            // split file name
+            $fileName = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+            // extenstion
+            $ext = $request->file('image')->getClientOriginalExtension();
+            // filename
+            $fileNameStore = $fileName.'_'.time().'.'.$ext;
+            // add image
+            $path = $request->file('image')->storeAs('public/images',$fileNameStore);
+
+        }else{
+            $fileName = 'null';
+        }
 
         // create article
         $article = new Article();
@@ -39,6 +69,7 @@ class ArticlesController extends Controller
         $article->user_id = Auth::user()->id;
         $article->description = $request->input('desc');
         $article->category_id = $request->input('cat');
+        $article->image = $fileNameStore;
         $article->save();
 
         return redirect('/articles')->with('success', 'Articles Created');
@@ -47,12 +78,17 @@ class ArticlesController extends Controller
     public function show($id)
     {
         $article = Article::find($id);
+        // dd($article->category->name);
         return view('articles.show')->with('article',$article);
     }
 
     public function edit($id)
     {
         $article = Article::find($id);
+        // Check User or Admin
+        if(auth()->user()->id != $article->user_id && auth()->user()->role != 'Admin'){
+            return redirect('/articles')->with('error', 'Unauthorized');
+        }
         return view('articles.edit')->with('article',$article);
     }
 
@@ -63,11 +99,28 @@ class ArticlesController extends Controller
             'desc' => 'required',
         ]);
 
+        if($request->hasFile('image')){
+            // Get filename with extension
+            $fileNameWithExt = $request->file('image')->getClientOriginalName();
+            // split file name
+            $fileName = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+            // extenstion
+            $ext = $request->file('image')->getClientOriginalExtension();
+            // filename
+            $fileNameStore = $fileName.'_'.time().'.'.$ext;
+            // add image
+            $path = $request->file('image')->storeAs('public/images',$fileNameStore);
+
+        }
+
         // update article
         $article = Article::find($id);
         $article->title = $request->input('title');
         $article->description = $request->input('desc');
         $article->category_id = $request->input('cat');
+        if($request->hasFile('image')){
+            $article->image = $fileNameStore;
+        }
         $article->save();
 
         return redirect('/articles')->with('success', 'Articles Updated');
@@ -77,6 +130,14 @@ class ArticlesController extends Controller
     public function destroy($id)
     {
         $article = Article::find($id);
+        // dd($article);
+        // Check User or Admin
+        if(auth()->user()->id != $article->user_id || auth()->user()->role != 'Admin'){
+            return redirect('/articles')->with('error', 'Unauthorized');
+        }
+        if($article->image != 'null'){
+            Storage::delete('public/images/'.$article->image);
+        }
         $article->delete();
         return redirect('/articles')->with('success', 'Articles Deleted');
     }
